+++ tests/DevEngines/Fail.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Arokettu\Random\Tests\DevEngines;
+
+use Random\Engine;
+
+final class Fail implements Engine
+{
+    public function generate(): string
+    {
+        return '';
+    }
+}