+++ tests/DevEngines/Xorshift32.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Random\Engine;
+namespace Arokettu\Random\Tests\DevEngines;
 
 use Random\Engine;