+++ tests/DevEngines/Zeros.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Arokettu\Random\Tests\DevEngines;
+
+use Random\Engine;
+
+final class Zeros implements Engine
+{
+    public function generate(): string
+    {
+        return "\0\0\0\0";
+    }
+}